refactor: 각 페이지 하위에 헤더를 두어 합성 방식으로 컴포넌트를 조합하도록 변경
