+++ src/components/shoppingCart/ShoppingCartHeader/ShoppingCartHeader.styles.ts
