+++ src/components/shoppingCart/ShoppingCartHeader/ShoppingCartHeader.tsx
@@ -1,10 +1,13 @@
-import * as S from "./Header.styles";
+import * as S from "./ShoppingCartHeader.styles";
 interface HeaderProps {
   title: string;
   description?: string;
 }
 
-export default function Header({ title, description = "" }: HeaderProps) {
+export default function ShoppingCartHeader({
+  title,
+  description = "",
+}: HeaderProps) {
   return (
     <S.Section>
       <S.Title>{title}</S.Title>
